①更新至1.21.3

+++ src/main/java/cn/anecansaitin/freecameraapi/CameraModifierManager.java
@@ -640,7 +640,7 @@ public class CameraModifierManager {
             fov = FOV;
         }
 
-        event.setFOV(fov);
+        event.setFOV((float) fov);
     }
 
     @SubscribeEvent
